pyproject.toml
[project]
name = "transcription-vae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Christian Ramirez", email = "christian.ramirez1@umassmed.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "dask>=2025.2.0",
    "ipython>=8.34.0",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.6",
    "keras>=3.9.0",
    "matplotlib>=3.10.1",
    "narwhals>=1.30.0",
    "networkx>=3.4.2",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pybigtools>=0.2.4",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "swifter>=1.4.0",
    "sympy>=1.13.3",
    "tensorflow>=2.19.0",
    "torch>=2.4.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
